Added gitignore files, fixed some warnings

+++ riscv-trace/Source/CodeParser.cpp
@@ -53,8 +53,8 @@ TCodeParser::CCode TCodeParser::getCodeFromFile(const string &code, map<string, 
                 }
             }
         }
-        symbNum += buf.length() + 1;
-        lineNum += count(buf.begin(), buf.end(), '\n');
+        symbNum += (int) buf.length() + 1;
+        lineNum += (int) count(buf.begin(), buf.end(), '\n');
     }
     return result;
 }
